requirements.txt
flask==2.3.3
easyocr==1.7.1
pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
flask-cors==4.0.0
sqlalchemy==2.0.20
firebase-admin==6.2.0
pyrebase4==4.7.1
python-jwt==4.0.0
requests==2.31.0 